+++ react/components/Common/CurrentStock.tsx
