+++ react/components/Common/toggle.tsx
@@ -2,7 +2,7 @@ import React, { FC, useState } from 'react'
 import { useMutation, useQuery } from 'react-apollo'
 import { Toggle } from 'vtex.styleguide'
 
-import { DividerArea } from '../Divider/divider'
+import { DividerArea } from './divider'
 import setTaxConfigurationMutation from '../../queries/setTaxConfiguration.gql'
 import getTaxConfiguration from '../../queries/getTaxConfiguration.gql'
